requirements.txt
Django==1.6.1
Pillow==2.6.1
South==1.0.1
argparse==1.2.1
django-betterforms==1.1.1
django-stdimage==1.1.2
progressbar==2.2
six==1.8.0
wsgiref==0.1.2
